[refactor]: 디렉터리 구조 변경
softeer5th#193
