+++ backend/src/main/java/com/ddbb/dingdong/application/usecase/user/EnrollFCMTokenUseCase.java
@@ -3,8 +3,8 @@ package com.ddbb.dingdong.application.usecase.user;
 import com.ddbb.dingdong.application.common.Params;
 import com.ddbb.dingdong.application.common.UseCase;
 import com.ddbb.dingdong.domain.common.exception.DomainException;
-import com.ddbb.dingdong.infrastructure.firebase.FCMTokenManagement;
-import com.ddbb.dingdong.infrastructure.firebase.error.FCMException;
+import com.ddbb.dingdong.infrastructure.notification.firebase.FCMTokenManagement;
+import com.ddbb.dingdong.infrastructure.notification.firebase.error.FCMException;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
